LayoutTests/fast/regions/auto-height/single-auto-height-region-no-breaks.html

<!doctype html>
<html>
    <head>
        <!--
        Test for regions auto height prototype. A single region auto-height catches both p's.
        -->
        <script src="../../../fast/js/resources/js-test-pre.js"></script>
        <style>
            #article { -webkit-flow-into: flow; }
            #region { -webkit-flow-from: flow; border: 1px solid red; width: 300px; }
        </style>
    </head>
    <body>
        <div id="article">
            <p id="p1">Text inside p1 that flows into region. Text inside p1 that flows into region. </p>
            <p id="p2">Text inside p2 that flows into region. Text inside p2 that flows into region. </p>
        </div>
        <div id="region"></region>
        <script>
            if (window.layoutTestController)
                window.layoutTestController.dumpAsText();

            description("Test for regions auto-height prototype");
            var namedFlow = document.webkitGetFlowByName("flow");
            var p1RegionList = namedFlow.getRegionsByContentNode(document.getElementById("p1"));
            var p2RegionList = namedFlow.getRegionsByContentNode(document.getElementById("p2"));
            shouldEvaluateTo("p1RegionList.length", 1);
            shouldEvaluateTo("p2RegionList.length", 1);
            shouldBeEqualToString("p1RegionList.item(0).id", "region");
            shouldBeEqualToString("p2RegionList.item(0).id", "region");

            document.getElementById("article").style.visibility = "hidden";
            document.getElementById("region").style.visibility = "hidden";
        </script>
    </body>
    <script src="../../../fast/js/resources/js-test-post.js"></script>
</html>
