LayoutTests/fast/regions/auto-height/single-auto-height-region-no-breaks-expected.txt

Test for regions auto-height prototype

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS p1RegionList.length is 1
PASS p2RegionList.length is 1
PASS p1RegionList.item(0).id is "region"
PASS p2RegionList.item(0).id is "region"
PASS successfullyParsed is true

TEST COMPLETE
